旧大陆（国旗尚未制作）

+++ py files/v2_parser_py/V2_parser.py
@@ -15,20 +15,21 @@ class V2_parser:
     
     def process(self, proccls):
         for fileObj in self.fileList:
-            print "fucking " + fileObj
             f = open(fileObj,'r+')
             all_the_lines=f.readlines()
             f.seek(0)
             
             v2 = V2()
             v2.load(all_the_lines)
+            v = v2.struct
             v2.struct = proccls.process(v2.struct)
             save_lines = v2.save()
-            
-            f.truncate()
-            for line in save_lines:
-                f.write(line)
-            f.close()  
+            if not (v2 is v):
+                print "fucking " + fileObj
+                f.truncate()
+                for line in save_lines:
+                    f.write(line)
+                f.close()  
             
     def processWithFilename(self, proccls):
         for fileObj in self.fileList:
